LA-436 Update test resource generation process
`deploy-rpc-playbooks.sh` changes:
- calling `generate-resources.yml` is added to allow the resources to be
  set up before Telegraf is deployed.
- the rpc-maas var `telegraf_plugin_ping_urls` is set with the floating
  IP addresses of the generated resources so they can be monitored
during the upgrade.

`generate-resources.yml` changes:
- to accommodate the requirement to run this playbook in
  `deploy-rpc-playbook.sh` it is moved to `rpcd/playbooks`
- `os-tempest-install.yml` added to removing the need for rpc-gating to
  manage this requirement.
- checks for existing instances created by this playbook and skips if
  there are any given that they only need to be created at the beginning
of an upgrade.

`generate_resources.py` changes:
- updated to make use of the OpenStack Python SDK, this simplifies the
  code
- floating IP addresses are now assigned to the created servers
- a security group is created to ensure the servers can be pinged

`maas_release` is updated to a SHA that includes support for the
Telegraf ping plugin.
